# ACCServer/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shadowDemo restful API.",
    "version": "1.0"
  },
  "host": "localhost:8030",
  "basePath": "/acc",
  "paths": {
...
    "/progConfig": {
      "get": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "progConfig"
        ],
        "summary": "详情;",
        "operationId": "getProgConfig",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getProgConfigResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "progConfig"
        ],
        "summary": "删除;",
        "operationId": "deleteProgConfig",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "progConfig"
        ],
        "summary": "查询列表;",
        "operationId": "searchProgConfig",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CreateEndTime": {
                  "description": "创建截止时间",
                  "type": "string",
                  "format": "date-time"
                },
                "CreateStartTime": {
                  "description": "创建开始时间",
                  "type": "string",
                  "format": "date-time"
                },
                "IDS": {
                  "description": "id",
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "PageNum": {
                  "type": "integer",
                  "format": "int64"
                },
                "PageSize": {
                  "type": "integer",
                  "format": "int64"
                },
                "ParamName": {
                  "description": "参数名称",
                  "type": "string"
                },
                "Type": {
                  "description": "类型",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/searchProgConfigResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
...
    "/user": {
      "get": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "user"
        ],
        "summary": "详情;",
        "operationId": "getUser",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getUserResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "user"
        ],
        "summary": "修改;",
        "operationId": "updateUser",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "user"
        ],
        "summary": "创建;",
        "operationId": "createUser",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "user"
        ],
        "summary": "删除;",
        "operationId": "deleteUser",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/genericSuccess"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    },
    "/users": {
      "post": {
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": [
              "read",
              "write"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "user"
        ],
        "summary": "查询列表;",
        "operationId": "searchUser",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "CreateEndTime": {
                  "description": "创建截止时间",
                  "type": "string",
                  "format": "date-time"
                },
                "CreateStartTime": {
                  "description": "创建开始时间",
                  "type": "string",
                  "format": "date-time"
                },
                "IDS": {
                  "description": "id",
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                },
                "PageNum": {
                  "type": "integer",
                  "format": "int64"
                },
                "PageSize": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/searchUserResponse"
          },
          "default": {
            "$ref": "#/responses/genericError"
          }
        }
      }
    }
  },
  "definitions": {
...
    "AccountState": {
      "type": "integer",
      "format": "int64",
      "x-go-package": "shadowDemo/zframework/model"
    },
...
    "GenericMessageBody": {
      "type": "object",
      "properties": {
        "Code": {
          "type": "integer",
          "format": "int64"
        },
        "Msg": {
          "type": "string"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
    "GetProgConfigResponseBody": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/ProgConfig"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "GetUserResponseBody": {
      "type": "object",
      "properties": {
        "Result": {
          "$ref": "#/definitions/User"
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
    "ProgConfig": {
      "type": "object",
      "properties": {
        "Comment": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "description": "创建人",
          "type": "string"
        },
        "Disabled": {
          "type": "boolean"
        },
        "Encrypted": {
          "type": "boolean"
        },
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "ParamName": {
          "type": "string"
        },
        "ParamValue": {
          "type": "string"
        },
        "Type": {
          "type": "string"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "shadowDemo/model/do"
    },
...
    "SearchProgConfigResponseBody": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProgConfig"
          }
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
    "SearchUserResponseBody": {
      "type": "object",
      "properties": {
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "Result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      },
      "x-go-package": "shadowDemo/ACCServer/router/vo"
    },
...
    "User": {
      "description": "sagger:model",
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "CreatedBy": {
          "type": "string"
        },
        "Creator": {
          "description": "来源终端(PC/后台注册/导入玩家/手机端/手机端H5/手机端ANDROID/手机端IOS)\n创建人ID",
          "type": "integer",
          "format": "int64"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "int64"
        },
        "LastLogUrl": {
          "description": "最后登录的访问链接",
          "type": "string"
        },
        "LastLoginAt": {
          "description": "最后登录时间",
          "type": "string",
          "format": "date-time"
        },
        "LastLoginIP": {
          "description": "最后登录IP",
          "type": "string"
        },
        "LastLoginIPAddr": {
          "description": "最后登录ip地址",
          "type": "string"
        },
        "Recommend": {
          "description": "设备信息\n钱包\nWallet do.Wallet `gorm:\"ForeignKey:WalletID\" gorm:\"association_foreignkey:UserID\" json:\"wallet\"`",
          "type": "string"
        },
        "Sca": {
          "type": "integer",
          "format": "int64"
        },
        "State": {
          "$ref": "#/definitions/AccountState"
        },
        "StateTime": {
          "description": "帐号状态,  正常/锁定/冻结/停用\n状态修改时间",
          "type": "string",
          "format": "date-time"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "WalletID": {
          "description": "钱包ID",
          "type": "integer",
          "format": "int64"
        },
        "account": {
          "description": "帐号",
          "type": "string",
          "x-go-name": "Account"
        },
        "address": {
          "description": "住址",
          "type": "string",
          "x-go-name": "Address"
        },
        "birthday": {
          "description": "生日",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Birthday"
        },
        "country": {
          "description": "性别\n国家地区",
          "type": "string",
          "x-go-name": "Country"
        },
        "email": {
          "description": "邮件",
          "type": "string",
          "x-go-name": "Email"
        },
        "loginPassword": {
          "description": "登录密码",
          "type": "string",
          "x-go-name": "LoginPassword"
        },
        "loginPasswordMd5": {
          "description": "登录密码的MD5加密字符串",
          "type": "string",
          "x-go-name": "LoginPasswordMd5"
        },
        "lsSecurePwdSet": {
          "description": "是否设置安全密码",
          "type": "boolean",
          "x-go-name": "IsSecurePwdSet"
        },
        "mobile": {
          "description": "手机",
          "type": "string",
          "x-go-name": "Mobile"
        },
        "nickname": {
          "description": "真实姓名",
          "type": "string",
          "x-go-name": "NickName"
        },
        "phoneCode": {
          "description": "默认地址",
          "type": "string",
          "x-go-name": "PhoneCode"
        },
        "qq": {
          "description": "Qq",
          "type": "string",
          "x-go-name": "Qq"
        },
        "recommanderPCode": {
          "description": "推荐人的推广码, (如果玩家是被推荐的）",
          "type": "string",
          "x-go-name": "RecommanderPCode"
        },
        "registIP": {
          "description": "注册IP",
          "type": "string",
          "x-go-name": "RegistIP"
        },
        "registURL": {
          "description": "注册URL",
          "type": "string",
          "x-go-name": "RegistURL"
        },
        "securePassword": {
          "description": "安全密码",
          "type": "string",
          "x-go-name": "SecurePassword"
        },
        "skype": {
          "description": "skype",
          "type": "string",
          "x-go-name": "Skype"
        },
        "timeZone": {
          "description": "时区",
          "type": "string",
          "x-go-name": "TimeZone"
        },
        "userAgentLevelID": {
          "description": "代理层级ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserAgentLevelID"
        },
        "userLevelID": {
          "description": "层级ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserLevelID"
        },
        "wechat": {
          "description": "微信",
          "type": "string",
          "x-go-name": "Wechat"
        }
      },
      "x-go-package": "shadowDemo/model/do"
    }
  },
  "responses": {
...
    "genericError": {
      "description": "错误消息",
      "schema": {
        "$ref": "#/definitions/GenericMessageBody"
      }
    },
    "genericSuccess": {
      "description": "成功消息",
      "schema": {
        "$ref": "#/definitions/GenericMessageBody"
      }
    },
...
    "getProgConfigResponse": {
      "schema": {
        "$ref": "#/definitions/GetProgConfigResponseBody"
      }
    },
    "getUserResponse": {
      "description": "详情出参",
      "schema": {
        "$ref": "#/definitions/GetUserResponseBody"
      }
    },
...
    "searchProgConfigResponse": {
      "description": "活动列表",
      "schema": {
        "$ref": "#/definitions/SearchProgConfigResponseBody"
      }
    },
    "searchUserResponse": {
      "description": "查询列表",
      "schema": {
        "$ref": "#/definitions/SearchUserResponseBody"
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}